// src/models/CityResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bus Ticket Booking API
 * API for managing bus ticket booking application
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';

/**
 * Response object for city details
 * @export
 * @interface CityResponse
 */
export interface CityResponse {
  /**
   * Name of the city
   * @type {string}
   * @memberof CityResponse
   */
  name: string;
}

/**
 * Check if a given object implements the CityResponse interface.
 */
export function instanceOfCityResponse(value: object): value is CityResponse {
  if (!('name' in value) || value['name'] === undefined) return false;
  return true;
}

export function CityResponseFromJSON(json: any): CityResponse {
  return CityResponseFromJSONTyped(json, false);
}

export function CityResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CityResponse {
  if (json == null) {
    return json;
  }
  return {
    name: json['name'],
  };
}

export function CityResponseToJSON(json: any): CityResponse {
  return CityResponseToJSONTyped(json, false);
}

export function CityResponseToJSONTyped(
  value?: CityResponse | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    name: value['name'],
  };
}